adding confusion matrix and a number of metrics associated with it
