+++ .travis.yml
@@ -3,7 +3,6 @@
 language: julia
 os:
   - linux
-  - osx
 julia:
 #  - 0.7
   - 1.0
