+++ src/measures/loss_functions_interface.jl
